requirements.txt
# Core Dependencies
ccxt>=4.0.0
pandas>=2.0.0
numpy>=1.24.0

# Technical Analysis
TA-Lib>=0.4.28
pandas-ta>=0.3.14b

# Machine Learning
scikit-learn>=1.3.0
xgboost>=2.0.0
tensorflow>=2.15.0
keras>=3.0.0

# Scheduling & Async
APScheduler>=3.10.0
asyncio>=3.4.3

# Notifications
python-telegram-bot>=20.0
requests>=2.31.0

# Configuration & Environment
python-dotenv>=1.0.0
pyyaml>=6.0

# Logging & Monitoring
colorlog>=6.7.0
Flask>=3.0.0

# Data Processing
scipy>=1.11.0
matplotlib>=3.8.0
seaborn>=0.13.0

# Testing (optional)
pytest>=7.4.0
pytest-asyncio>=0.21.0
